Leave the Wintab context open while Blender is running.
This change removes Wintab teardown and setup when switching between the
Win32 Pointer and Wintab APIs and instead handles it by enabling/disabling
the Wintab context if it exists.

When display layout changes Wintab still needs to be updated. Instead of
tearing the context down and setting it back up, update to the new out
coordinate mapping.

Signed-off-by: Nicholas Rishel <rishel.nick@gmail.com>